redesigned dashboard
