+++ Sweetsy/MealItemDetailView.swift
@@ -1,0 +1,19 @@
+//
+//  MealItemDetailView.swift
+//  Sweetsy
+//
+//  Created by Chetan Dhowlaghar on 5/29/24.
+//
+
+import SwiftUI
+
+struct MealItemDetailView: View {
+    var body: some View {
+        VStack {
+        }
+    }
+}
+
+#Preview {
+    MealItemDetailView()
+}
